requirements.txt
yt-dlp
mutagen
python-dotenv
requests
logging